Docks init
