+++ Cargo.toml
@@ -31,6 +31,7 @@ rand = "0.8"
 anyhow = "1.0"
 egui-notify = "0.15.0"
 once_cell = "1.19.0"
+egui_dock = "0.13.0"
 
 [target.'cfg(windows)'.dependencies]
 windows = { version = "0.58.0", features = [
